syslog + fcntl: extract into module/syslog/ + module/fcntl/
importing.rs shrinks by 585 lines.
