+++ pyre/pyre-interpreter/src/importing.rs
@@ -138,9 +138,9 @@ pub fn install_builtin_modules() {
     #[cfg(unix)]
     register_builtin_module("resource", init_resource);
     #[cfg(unix)]
-    register_builtin_module("fcntl", init_fcntl);
+    register_builtin_module("fcntl", crate::module::fcntl::moduledef::init);
     #[cfg(unix)]
-    register_builtin_module("syslog", init_syslog);
+    register_builtin_module("syslog", crate::module::syslog::moduledef::init);
     register_builtin_module("select", crate::module::select::moduledef::init);
     register_builtin_module("termios", crate::module::termios::moduledef::init);
     register_builtin_module("_socket", crate::module::_socket::moduledef::init);
@@ -1921,593 +1921,6 @@ fn init_resource(ns: &mut DictStorage) {
             pyre_object::w_int_new(libc::RLIM_INFINITY as i64),
         );
     }
-}
-
-/// fcntl module — PyPy: pypy/module/fcntl/interp_fcntl.py.
-///
-/// fcntl(fd, cmd, arg=0) / ioctl(fd, request, arg=0) / flock(fd, op) /
-/// lockf(fd, cmd, len=0, start=0, whence=0).  Backed by
-/// `rustpython_host_env::fcntl`.  Only the integer-argument forms are
-/// implemented; bytes-buffer (out-arg) variants are out of scope.
-fn init_fcntl(ns: &mut DictStorage) {
-    crate::dict_storage_store(
-        ns,
-        "fcntl",
-        crate::make_builtin_function("fcntl", |args| {
-            #[cfg(all(unix, feature = "host_env"))]
-            {
-                if args.len() < 2 {
-                    return Err(crate::PyError::type_error(
-                        "fcntl() requires at least 2 arguments",
-                    ));
-                }
-                if !unsafe { pyre_object::is_int(args[0]) }
-                    || !unsafe { pyre_object::is_int(args[1]) }
-                    || (args.len() >= 3 && !unsafe { pyre_object::is_int(args[2]) })
-                {
-                    return Err(crate::PyError::type_error(
-                        "fcntl() arguments must be integers",
-                    ));
-                }
-                let fd = (unsafe { pyre_object::w_int_get_value(args[0]) }) as i32;
-                let cmd = (unsafe { pyre_object::w_int_get_value(args[1]) }) as i32;
-                let arg = if args.len() >= 3 {
-                    unsafe { pyre_object::w_int_get_value(args[2]) as i32 }
-                } else {
-                    0
-                };
-                match rustpython_host_env::fcntl::fcntl_int(fd, cmd, arg) {
-                    Ok(v) => Ok(pyre_object::w_int_new(v as i64)),
-                    Err(e) => Err(crate::PyError::os_error_with_errno(
-                        e.raw_os_error().unwrap_or(0),
-                        format!("fcntl: {e}"),
-                    )),
-                }
-            }
-            #[cfg(not(all(unix, feature = "host_env")))]
-            {
-                let _ = args;
-                Err(crate::PyError::not_implemented(
-                    "fcntl.fcntl requires host_env feature",
-                ))
-            }
-        }),
-    );
-    crate::dict_storage_store(
-        ns,
-        "ioctl",
-        crate::make_builtin_function("ioctl", |args| {
-            #[cfg(all(unix, feature = "host_env"))]
-            {
-                if args.len() < 2 {
-                    return Err(crate::PyError::type_error(
-                        "ioctl() requires at least 2 arguments",
-                    ));
-                }
-                if !unsafe { pyre_object::is_int(args[0]) }
-                    || !unsafe { pyre_object::is_int(args[1]) }
-                    || (args.len() >= 3 && !unsafe { pyre_object::is_int(args[2]) })
-                {
-                    return Err(crate::PyError::type_error(
-                        "ioctl() arguments must be integers",
-                    ));
-                }
-                let fd = (unsafe { pyre_object::w_int_get_value(args[0]) }) as i32;
-                let raw_req = (unsafe { pyre_object::w_int_get_value(args[1]) }) as i64;
-                let request = rustpython_host_env::fcntl::normalize_ioctl_request(raw_req);
-                let arg = if args.len() >= 3 {
-                    unsafe { pyre_object::w_int_get_value(args[2]) as i32 }
-                } else {
-                    0
-                };
-                match rustpython_host_env::fcntl::ioctl_int(fd, request, arg) {
-                    Ok(v) => Ok(pyre_object::w_int_new(v as i64)),
-                    Err(e) => Err(crate::PyError::os_error_with_errno(
-                        e.raw_os_error().unwrap_or(0),
-                        format!("ioctl: {e}"),
-                    )),
-                }
-            }
-            #[cfg(not(all(unix, feature = "host_env")))]
-            {
-                let _ = args;
-                Err(crate::PyError::not_implemented(
-                    "fcntl.ioctl requires host_env feature",
-                ))
-            }
-        }),
-    );
-    crate::dict_storage_store(
-        ns,
-        "flock",
-        crate::make_builtin_function_with_arity(
-            "flock",
-            |args| {
-                #[cfg(all(unix, feature = "host_env"))]
-                {
-                    if args.len() < 2 {
-                        return Err(crate::PyError::type_error("flock() requires 2 arguments"));
-                    }
-                    if !unsafe { pyre_object::is_int(args[0]) }
-                        || !unsafe { pyre_object::is_int(args[1]) }
-                    {
-                        return Err(crate::PyError::type_error(
-                            "flock() arguments must be integers",
-                        ));
-                    }
-                    let fd = (unsafe { pyre_object::w_int_get_value(args[0]) }) as i32;
-                    let op = (unsafe { pyre_object::w_int_get_value(args[1]) }) as i32;
-                    match rustpython_host_env::fcntl::flock(fd, op) {
-                        Ok(_) => Ok(pyre_object::w_none()),
-                        Err(e) => Err(crate::PyError::os_error_with_errno(
-                            e.raw_os_error().unwrap_or(0),
-                            format!("flock: {e}"),
-                        )),
-                    }
-                }
-                #[cfg(not(all(unix, feature = "host_env")))]
-                {
-                    let _ = args;
-                    Err(crate::PyError::not_implemented(
-                        "fcntl.flock requires host_env feature",
-                    ))
-                }
-            },
-            2,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "lockf",
-        crate::make_builtin_function("lockf", |args| {
-            #[cfg(all(unix, feature = "host_env"))]
-            {
-                if args.len() < 2 {
-                    return Err(crate::PyError::type_error(
-                        "lockf() requires at least 2 arguments",
-                    ));
-                }
-                for (i, &a) in args.iter().enumerate().take(5) {
-                    if !unsafe { pyre_object::is_int(a) } {
-                        let _ = i;
-                        return Err(crate::PyError::type_error(
-                            "lockf() arguments must be integers",
-                        ));
-                    }
-                }
-                let fd = (unsafe { pyre_object::w_int_get_value(args[0]) }) as i32;
-                let cmd = (unsafe { pyre_object::w_int_get_value(args[1]) }) as i32;
-                let len = if args.len() >= 3 {
-                    unsafe { pyre_object::w_int_get_value(args[2]) }
-                } else {
-                    0
-                };
-                let start = if args.len() >= 4 {
-                    unsafe { pyre_object::w_int_get_value(args[3]) }
-                } else {
-                    0
-                };
-                let whence = if args.len() >= 5 {
-                    unsafe { pyre_object::w_int_get_value(args[4]) as i32 }
-                } else {
-                    0
-                };
-                match rustpython_host_env::fcntl::lockf(fd, cmd, len, start, whence) {
-                    Ok(v) => Ok(pyre_object::w_int_new(v as i64)),
-                    Err(rustpython_host_env::fcntl::LockfError::InvalidCmd) => {
-                        Err(crate::PyError::value_error("lockf: invalid cmd"))
-                    }
-                    Err(rustpython_host_env::fcntl::LockfError::Overflow(s)) => {
-                        Err(crate::PyError::value_error(format!("lockf: overflow: {s}")))
-                    }
-                    Err(rustpython_host_env::fcntl::LockfError::Io(e)) => {
-                        Err(crate::PyError::os_error_with_errno(
-                            e.raw_os_error().unwrap_or(0),
-                            format!("lockf: {e}"),
-                        ))
-                    }
-                }
-            }
-            #[cfg(not(all(unix, feature = "host_env")))]
-            {
-                let _ = args;
-                Err(crate::PyError::not_implemented(
-                    "fcntl.lockf requires host_env feature",
-                ))
-            }
-        }),
-    );
-    // `interp_fcntl.py:25-37 constant_names` — POSIX subset always
-    // exposed; Linux-specific block gated below.  I_* (System V
-    // STREAMS) are listed by PyPy but `if value is not None` filters
-    // them out at platform.configure time on every supported platform;
-    // not exposed here.
-    #[cfg(unix)]
-    {
-        macro_rules! cst {
-            ($name:literal, $val:expr) => {
-                crate::dict_storage_store(ns, $name, pyre_object::w_int_new($val as i64));
-            };
-        }
-        cst!("F_GETFD", libc::F_GETFD);
-        cst!("F_SETFD", libc::F_SETFD);
-        cst!("F_GETFL", libc::F_GETFL);
-        cst!("F_SETFL", libc::F_SETFL);
-        cst!("F_DUPFD", libc::F_DUPFD);
-        cst!("F_DUPFD_CLOEXEC", libc::F_DUPFD_CLOEXEC);
-        cst!("F_GETLK", libc::F_GETLK);
-        cst!("F_SETLK", libc::F_SETLK);
-        cst!("F_SETLKW", libc::F_SETLKW);
-        cst!("F_GETOWN", libc::F_GETOWN);
-        cst!("F_SETOWN", libc::F_SETOWN);
-        cst!("F_RDLCK", libc::F_RDLCK);
-        cst!("F_WRLCK", libc::F_WRLCK);
-        cst!("F_UNLCK", libc::F_UNLCK);
-        cst!("FD_CLOEXEC", libc::FD_CLOEXEC);
-        cst!("LOCK_SH", libc::LOCK_SH);
-        cst!("LOCK_EX", libc::LOCK_EX);
-        cst!("LOCK_UN", libc::LOCK_UN);
-        cst!("LOCK_NB", libc::LOCK_NB);
-
-        // Linux-only fcntl constants.  Values for ones libc does not
-        // expose (F_GETSIG/F_SETSIG/F_GETLK64/F_SETLK64/F_SETLKW64/
-        // F_EXLCK/F_SHLCK/LOCK_MAND/LOCK_READ/LOCK_WRITE/LOCK_RW/DN_*)
-        // come straight from Linux <fcntl.h>, matching the hardcoded
-        // overrides at `interp_fcntl.py:48-52`.
-        #[cfg(target_os = "linux")]
-        {
-            cst!("F_SETLEASE", libc::F_SETLEASE);
-            cst!("F_GETLEASE", libc::F_GETLEASE);
-            cst!("F_NOTIFY", libc::F_NOTIFY);
-            cst!("F_GETSIG", 11);
-            cst!("F_SETSIG", 10);
-            cst!("F_GETLK64", 12);
-            cst!("F_SETLK64", 13);
-            cst!("F_SETLKW64", 14);
-            cst!("F_EXLCK", 4);
-            cst!("F_SHLCK", 8);
-            cst!("LOCK_MAND", 32);
-            cst!("LOCK_READ", 64);
-            cst!("LOCK_WRITE", 128);
-            cst!("LOCK_RW", 192);
-            cst!("DN_ACCESS", 1);
-            cst!("DN_MODIFY", 2);
-            cst!("DN_CREATE", 4);
-            cst!("DN_DELETE", 8);
-            cst!("DN_RENAME", 16);
-            cst!("DN_ATTRIB", 32);
-            cst!("DN_MULTISHOT", 0x80000000u32);
-            cst!("F_ADD_SEALS", libc::F_ADD_SEALS);
-            cst!("F_GET_SEALS", libc::F_GET_SEALS);
-            cst!("F_SEAL_SEAL", libc::F_SEAL_SEAL);
-            cst!("F_SEAL_SHRINK", libc::F_SEAL_SHRINK);
-            cst!("F_SEAL_GROW", libc::F_SEAL_GROW);
-            cst!("F_SEAL_WRITE", libc::F_SEAL_WRITE);
-            cst!("F_SETPIPE_SZ", libc::F_SETPIPE_SZ);
-            cst!("F_GETPIPE_SZ", libc::F_GETPIPE_SZ);
-        }
-    }
-}
-
-/// syslog module — PyPy: pypy/module/syslog/interp_syslog.py.
-///
-/// openlog / syslog / closelog / setlogmask.  Backed by
-/// `rustpython_host_env::syslog`.  Unix-only.
-fn init_syslog(ns: &mut DictStorage) {
-    crate::dict_storage_store(
-        ns,
-        "openlog",
-        crate::make_builtin_function("openlog", |args| {
-            #[cfg(all(unix, feature = "host_env"))]
-            {
-                let ident = args.first().and_then(|&a| unsafe {
-                    if pyre_object::is_str(a) {
-                        std::ffi::CString::new(pyre_object::w_str_get_value(a))
-                            .ok()
-                            .map(|c| c.into_boxed_c_str())
-                    } else {
-                        None
-                    }
-                });
-                if args
-                    .iter()
-                    .skip(1)
-                    .any(|&a| !unsafe { pyre_object::is_int(a) })
-                {
-                    return Err(crate::PyError::type_error(
-                        "openlog(): logoption and facility must be integers",
-                    ));
-                }
-                let logoption = args
-                    .get(1)
-                    .map(|&a| unsafe { pyre_object::w_int_get_value(a) } as i32)
-                    .unwrap_or(0);
-                let facility = args
-                    .get(2)
-                    .map(|&a| unsafe { pyre_object::w_int_get_value(a) } as i32)
-                    .unwrap_or(libc::LOG_USER);
-                rustpython_host_env::syslog::openlog(ident, logoption, facility);
-                Ok(pyre_object::w_none())
-            }
-            #[cfg(not(all(unix, feature = "host_env")))]
-            {
-                let _ = args;
-                Err(crate::PyError::not_implemented(
-                    "syslog.openlog requires host_env feature",
-                ))
-            }
-        }),
-    );
-    crate::dict_storage_store(
-        ns,
-        "syslog",
-        crate::make_builtin_function("syslog", |args| {
-            #[cfg(all(unix, feature = "host_env"))]
-            {
-                let (priority, msg_obj) = if args.len() >= 2 {
-                    if !unsafe { pyre_object::is_int(args[0]) } {
-                        return Err(crate::PyError::type_error(
-                            "syslog(): priority must be an integer",
-                        ));
-                    }
-                    (
-                        unsafe { pyre_object::w_int_get_value(args[0]) as i32 },
-                        args[1],
-                    )
-                } else if args.len() == 1 {
-                    (libc::LOG_INFO, args[0])
-                } else {
-                    return Err(crate::PyError::type_error("syslog() requires a message"));
-                };
-                if !unsafe { pyre_object::is_str(msg_obj) } {
-                    return Err(crate::PyError::type_error(
-                        "syslog(): message must be a string",
-                    ));
-                }
-                let msg = unsafe { pyre_object::w_str_get_value(msg_obj) };
-                if let Ok(cmsg) = std::ffi::CString::new(msg) {
-                    rustpython_host_env::syslog::syslog(priority, &cmsg);
-                }
-                Ok(pyre_object::w_none())
-            }
-            #[cfg(not(all(unix, feature = "host_env")))]
-            {
-                let _ = args;
-                Err(crate::PyError::not_implemented(
-                    "syslog.syslog requires host_env feature",
-                ))
-            }
-        }),
-    );
-    crate::dict_storage_store(
-        ns,
-        "closelog",
-        crate::make_builtin_function_with_arity(
-            "closelog",
-            |_| {
-                #[cfg(all(unix, feature = "host_env"))]
-                {
-                    rustpython_host_env::syslog::closelog();
-                }
-                Ok(pyre_object::w_none())
-            },
-            0,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "setlogmask",
-        crate::make_builtin_function_with_arity(
-            "setlogmask",
-            |args| {
-                #[cfg(all(unix, feature = "host_env"))]
-                {
-                    let mask = if let Some(&a) = args.first() {
-                        if !unsafe { pyre_object::is_int(a) } {
-                            return Err(crate::PyError::type_error(
-                                "setlogmask(): argument must be an integer",
-                            ));
-                        }
-                        unsafe { pyre_object::w_int_get_value(a) as i32 }
-                    } else {
-                        return Err(crate::PyError::type_error("setlogmask() missing argument"));
-                    };
-                    return Ok(pyre_object::w_int_new(
-                        rustpython_host_env::syslog::setlogmask(mask) as i64,
-                    ));
-                }
-                #[cfg(not(all(unix, feature = "host_env")))]
-                {
-                    let _ = args;
-                    Err(crate::PyError::not_implemented(
-                        "syslog.setlogmask requires host_env feature",
-                    ))
-                }
-            },
-            1,
-        ),
-    );
-    // Priorities + facilities (POSIX subset matching CPython).
-    #[cfg(unix)]
-    {
-        crate::dict_storage_store(
-            ns,
-            "LOG_EMERG",
-            pyre_object::w_int_new(libc::LOG_EMERG as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_ALERT",
-            pyre_object::w_int_new(libc::LOG_ALERT as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_CRIT",
-            pyre_object::w_int_new(libc::LOG_CRIT as i64),
-        );
-        crate::dict_storage_store(ns, "LOG_ERR", pyre_object::w_int_new(libc::LOG_ERR as i64));
-        crate::dict_storage_store(
-            ns,
-            "LOG_WARNING",
-            pyre_object::w_int_new(libc::LOG_WARNING as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_NOTICE",
-            pyre_object::w_int_new(libc::LOG_NOTICE as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_INFO",
-            pyre_object::w_int_new(libc::LOG_INFO as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_DEBUG",
-            pyre_object::w_int_new(libc::LOG_DEBUG as i64),
-        );
-        crate::dict_storage_store(ns, "LOG_PID", pyre_object::w_int_new(libc::LOG_PID as i64));
-        crate::dict_storage_store(
-            ns,
-            "LOG_CONS",
-            pyre_object::w_int_new(libc::LOG_CONS as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_NDELAY",
-            pyre_object::w_int_new(libc::LOG_NDELAY as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_NOWAIT",
-            pyre_object::w_int_new(libc::LOG_NOWAIT as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_PERROR",
-            pyre_object::w_int_new(libc::LOG_PERROR as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_KERN",
-            pyre_object::w_int_new(libc::LOG_KERN as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_USER",
-            pyre_object::w_int_new(libc::LOG_USER as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_MAIL",
-            pyre_object::w_int_new(libc::LOG_MAIL as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_DAEMON",
-            pyre_object::w_int_new(libc::LOG_DAEMON as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_AUTH",
-            pyre_object::w_int_new(libc::LOG_AUTH as i64),
-        );
-        crate::dict_storage_store(ns, "LOG_LPR", pyre_object::w_int_new(libc::LOG_LPR as i64));
-        crate::dict_storage_store(
-            ns,
-            "LOG_NEWS",
-            pyre_object::w_int_new(libc::LOG_NEWS as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_UUCP",
-            pyre_object::w_int_new(libc::LOG_UUCP as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_CRON",
-            pyre_object::w_int_new(libc::LOG_CRON as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_SYSLOG",
-            pyre_object::w_int_new(libc::LOG_SYSLOG as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_LOCAL0",
-            pyre_object::w_int_new(libc::LOG_LOCAL0 as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_LOCAL1",
-            pyre_object::w_int_new(libc::LOG_LOCAL1 as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_LOCAL2",
-            pyre_object::w_int_new(libc::LOG_LOCAL2 as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_LOCAL3",
-            pyre_object::w_int_new(libc::LOG_LOCAL3 as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_LOCAL4",
-            pyre_object::w_int_new(libc::LOG_LOCAL4 as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_LOCAL5",
-            pyre_object::w_int_new(libc::LOG_LOCAL5 as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_LOCAL6",
-            pyre_object::w_int_new(libc::LOG_LOCAL6 as i64),
-        );
-        crate::dict_storage_store(
-            ns,
-            "LOG_LOCAL7",
-            pyre_object::w_int_new(libc::LOG_LOCAL7 as i64),
-        );
-    }
-    // `Modules/syslogmodule.c syslog_log_mask / syslog_log_upto` —
-    // helpers for building setlogmask() arguments.
-    //   LOG_MASK(pri)  → 1 << pri
-    //   LOG_UPTO(pri)  → (1 << (pri + 1)) - 1
-    crate::dict_storage_store(
-        ns,
-        "LOG_MASK",
-        crate::make_builtin_function_with_arity(
-            "LOG_MASK",
-            |args| {
-                let pri =
-                    crate::baseobjspace::int_w(args.first().copied().ok_or_else(|| {
-                        crate::PyError::type_error("LOG_MASK() missing argument")
-                    })?)?;
-                Ok(pyre_object::w_int_new(1i64 << pri))
-            },
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "LOG_UPTO",
-        crate::make_builtin_function_with_arity(
-            "LOG_UPTO",
-            |args| {
-                let pri =
-                    crate::baseobjspace::int_w(args.first().copied().ok_or_else(|| {
-                        crate::PyError::type_error("LOG_UPTO() missing argument")
-                    })?)?;
-                Ok(pyre_object::w_int_new((1i64 << (pri + 1)) - 1))
-            },
-            1,
-        ),
-    );
 }
 
 /// itertools stub
