+++ pyre/pyre-interpreter/src/module/fcntl/mod.rs
@@ -1,0 +1,8 @@
+//! fcntl module — PyPy: pypy/module/fcntl/
+//!
+//! fcntl(fd, cmd, arg=0) / ioctl(fd, request, arg=0) / flock(fd, op) /
+//! lockf(fd, cmd, len=0, start=0, whence=0).  Integer-argument forms
+//! only.
+
+pub mod interp_fcntl;
+pub mod moduledef;
