+++ pyre/pyre-interpreter/src/module/fcntl/moduledef.rs
@@ -1,0 +1,7 @@
+//! fcntl module definition — PyPy: pypy/module/fcntl/moduledef.py
+
+use crate::DictStorage;
+
+pub fn init(ns: &mut DictStorage) {
+    super::interp_fcntl::register_module(ns);
+}
