+++ pyre/pyre-interpreter/src/module/mod.rs
@@ -19,10 +19,12 @@ pub mod _sre;
 pub mod _weakref;
 pub mod atexit;
 pub mod faulthandler;
+pub mod fcntl;
 pub mod math;
 pub mod mmap;
 pub mod operator;
 pub mod select;
 pub mod sys;
+pub mod syslog;
 pub mod termios;
 pub mod time;
