+++ pyre/pyre-interpreter/src/module/syslog/mod.rs
@@ -1,0 +1,7 @@
+//! syslog module — PyPy: pypy/module/syslog/
+//!
+//! openlog / syslog / closelog / setlogmask backed by
+//! `rustpython_host_env::syslog`.  Unix-only.
+
+pub mod interp_syslog;
+pub mod moduledef;
